Revert "补偿修改"
This reverts commit 3eba090631c2ec0af24cc5af2dad74585db20296.

+++ src/main/java/com/rjxx/taxeasy/configuration/DirectAmqpConfiguration.java
@@ -79,9 +79,9 @@ public class DirectAmqpConfiguration {
 
                         }else if(null !=invoiceResponse && invoiceResponse.getReturnCode() == null){
                             InvoiceResponse invoiceResponse2 =  skService.SkServerKP(kplsh);
-                            /*if (null !=invoiceResponse2 && invoiceResponse2.getReturnCode().equals("0000")){
+                            if (null !=invoiceResponse2 && invoiceResponse2.getReturnCode().equals("0000")){
                                 kpcfService.deleteById(kplsh);
-                            }*/
+                            }
 
                         }
                     }
